playwright-python/requirements.txt
playwright==1.41.1
pytest==7.4.4
python-dotenv==1.0.1
